requirements.txt
Flask==0.10.1
Jinja2==2.6
Werkzeug==0.8.3
gunicorn==0.17.2
SQLAlchemy==0.9.7
psycopg2==2.5.3
python-postmark==0.4.3
hashlib==20081119
